refactor(cli): split command option metadata
