+++ .sampo/changesets/839-split-command-options.md
@@ -1,0 +1,5 @@
+---
+npm/wp-typia: patch
+---
+
+Split shared CLI command option metadata into command-focused modules while keeping parser and routing aggregation behavior centralized.
