+++ packages/wp-typia/src/command-options/templates.ts
@@ -1,0 +1,11 @@
+import type { CommandOptionMetadataMap } from './types';
+
+/**
+ * Shared `wp-typia templates` option metadata.
+ */
+export const TEMPLATES_OPTION_METADATA = {
+  id: {
+    description: 'Template id for `templates inspect`.',
+    type: 'string',
+  },
+} as const satisfies CommandOptionMetadataMap;
