+++ packages/wp-typia/src/command-options/types.ts
@@ -1,0 +1,35 @@
+export type CommandOptionMetadata = {
+  argumentKind?: 'flag';
+  description: string;
+  short?: string;
+  type: 'boolean' | 'string';
+};
+
+export type CommandOptionMetadataMap = Record<string, CommandOptionMetadata>;
+
+export type ParsedCommandArgv = {
+  flags: Record<string, unknown>;
+  positionals: string[];
+};
+
+export type ShortOptionDescriptor = {
+  name: string;
+  type: CommandOptionMetadata['type'];
+};
+
+export type CommandOptionParser = {
+  booleanOptionNames: Set<string>;
+  shortFlagMap: Map<string, ShortOptionDescriptor>;
+  stringOptionNames: Set<string>;
+};
+
+export type CommandOptionGroupName =
+  | 'global'
+  | 'create'
+  | 'add'
+  | 'init'
+  | 'migrate'
+  | 'mcp'
+  | 'sync'
+  | 'doctor'
+  | 'templates';
